LayoutTests/accessibility/mac/progressbar-indeterminate-expected.txt

X

This tests that elements with progressbar role return 0 for min and max value, if indeterminate.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS prog1.maxValue is 0
PASS prog1.minValue is 0
PASS prog2.maxValue is 0
PASS prog2.minValue is 0
PASS successfullyParsed is true

TEST COMPLETE
